use size_t to avoid compiler warning
git-svn-id: http://vtp.googlecode.com/svn/trunk@6875 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainApps/VTBuilder/vtImage.h
@@ -144,7 +144,7 @@ public:
 	GDALDataset *m_pDataset;
 	LineBufferGDAL m_linebuf;
 
-	int NumBitmaps() const { return m_Bitmaps.size(); }
+	size_t NumBitmaps() const { return m_Bitmaps.size(); }
 	BitmapInfo &GetBitmapInfo(int i) { return m_Bitmaps[i]; }
 	int NumBitmapsInMemory();
 	int NumBitmapsOnDisk();
